Renderer: implementation of the core architecture and basic shape primitives
Changed the renderer design from a command list on CPU (1 draw call per primitive)
to a command list on the fragment shader code.
This allows for nice looking pixel based effects based on procedural shape
definitions such as drop shadows (not implemented yet) and round corners without
having to tessellate the vertex list.

The renderer now draws a simple full size quad and passes a list of primitive
shape parameters to the shader. The list is built when the component that uses
the renderer calls functions like addRect() and cleared on every draw.
There is no change to the client API except for more primitives.

Supported Shapes:
- filled rectangle with possibly round corners
- rectangle outline, also with possibly round corners
- filled circle
- circle outline
- triangle
- line.

Removed concept of clear color from the API. Now it's hardcoded in the shader.
(could be changed to a uniform to make it configurable)

ToDo:
- line and triangle are currently rendering as rects.
- support a bigger number of shapes.

Closes #1