Renderer: have style commands as a separate palette instead of interleaved with shapes
Style (line, corner width and color) is now in its own list which is indexed
by every shape command, instead of having the style as state between shapes.
This allows shapes to be interpreted out of order, since they explicitly index
their style.

The renderer still does "state tracking" and pushes a new style if it is
different from the previous one without checking for duplication of styles
before that.