Timeline & Renderer: add 'views' for panning and zooming
A View is an area in pixels with its own transform (offset and scale).
Shapes are added to a view in the base coordinate space. The renderer
will transform the shapes according to the current zoom and pan and it
will clip shapes that fall on the boundary of the view area.

Renderer:
- Add support for 'views' which get pushed as state.
- When there is a view, shapes will get transformed by it.
- Scale and offset are indenpendent in X and Y. Uniform distances (corner
and line width) get scaled by the smallest independent scale.
TODO: maybe these shouldn't get scaled at all? Optional?
- Shapes get accurately clipped in the fragment shader. (new command)
- AddPrimitveShape can now return false and prevent emitting a shape.
- Pass viewport size to the shader, not just height.
- Remove unused MVP.

Timeline:
- Add an horizontal zoom area for the timeline range. Height is not affected.
- Custom view logic was needed to handle:
  - text rendering, because that is rendered by a separate 2D canvas.
  - playhead, which should have a pos set by the view, but not the geo.
  - playhead click interaction.
- Interaction controls are MMB to pan, scroll to zoom and a button to fit.
More to follow.