Renderer: add very basic implementation for a text primitive
New addText API renders a string in a monospace font with advance and size being
zoom independent. It adds a new GLYPH command per character which renders a
glyph from a primitive atlas to the command's rect.

Shadow and color are temporarily baked in, as well as the font family, weight,
size, etc. This is just the minimum necessary to see something on screen.

+ refactor of addPrimitiveShape for re-use by the text primitive.
+ minor comments cleanup.