Image Rendering: fix for standalone images + improvements to fallback textures
- Bugfix: rendering more than one single image would always render the first one (bundles were fine).
This was because of a missing +5 for the sampler index.

- Fallback texture improvements/refactoring:
The "fallback textures" concept was split. There is still a 1px fallback texture for 2D and array
images, but they are now only used as binding target for the unused samplers in a frame, they should
never show (green).
For "fallback" error cases, the images will draw in a shade of pink to draw attention, also in the
case when out of samplers to support different textures, which previously only had a console warning.
Images will show in grey while they are still loading.
End users might see grey rectangles as visual feedback while they wait, but pink or green clearly
indicates unexpected application behaviour.
The new error and loading "textures" don't spend sampler or a texture object resources :)

- JS->TS: added some missing type information and fixed a few complaints.