Assets: Turn Inconsolata font into blob
This removes the .png requirement from uirenderer-canvas, and provides
the font to the renderer as a blob url on initialization.

There is a tradeoff between simplicity of use and file size, and at
this moment we prefer to have a "no configuration" experience rather
than requiring for the font png file to be included at every use of
the uirenderer-canvas.

The png file is kept as a reference, and whenever it's updated, a new
base64 string needs to be generated and included in blobs.ts.