JS->TS: Correct addImage() signatures to accept null or undefined textures from the client.
Linting caught that the code was checking for a null texture case, but the signature did not allow for it.
It seems more convenient for the client (and a JS client) to not have to ensure a correct texture and the
renderer will gracefully render it as an error. Clients can then assign-load images on mount instead of
on construction without asserting that the texture will really be there because of lifetime.

+ little refactor for readability
+ assume there is always a View. It is not guaranteed, so it was right to complain, but I preferred to
"know what I'm doing" than to add checks for undefined everywhere.