Continuation of WIP Setup upgrade for development and packaging
This brings everything closer to working.
UI Renderer compiles with tsc and lints.
All together is not working due to an import error for the GLSL files. The docs don't run.

I don't think the package is being distributed as it should ideally.
Currently, it is distributing types and the original ts and glsl files, but no transpiled js.
Vite is a dev dependency, but it is not being used to bundle the package for distribution.
Currently the dependent packages are forced to use vite and vite-glsl, which is not the intention.

The docs submodule is making things complicated.
It makes for a combinatorial explosion of dev dependencies and of thinking of the inheritance of configuration.

Not done:
- Did not review or fix the actual TS error and linting rules.
- Package the renderer for distribution with js and the shaders inlined.
- Double check if 'types' in TS config correct? and if the extra vue vite files in the docs could be deleted.

Done:

Improved TS configuration.
- Reviewed options for compilation and module resolution.
- tsconfig.json now applies to all project files, including the blob.ts assets, but not the config files or glsl.

Fixed ESLint setup.
- ESLint no longer needs the FlatCompat workaround for the new flat config way of configuration. Removed it.
- Reworked the configuration objects, but I'm not fully happy with it.
The problem is that typescript-eslint recommends simply plonking `eslint_ts.configs.recommended` as the full object (for both js and ts),
but the typescript configurations are made to apply to all files (including js) and don't come with a plugin and parser, so as soon as
changing recommended to recommendedTypeChecked, the eslint config file will error linting itself (oh the irony).
This has costed me too many hours by now, so it'll stay as is. I made objects that extend recommended rules and add file filters.
It seems like a lot of time configuring rules for files that should possibly not exist just to have a nice fallback.
- Made vite.config a js instead of ts file because I don't want to spend more time trying to lint it but not compiling it for building.
- Removed eslintConfig from the package.json, now it's only eslint.config.js.
- There is no difference between `eslint` and `eslint .` experimentally and according to the docs when using flatconfig.
- eslint --inspect-config is a life saver.

Added missing dev dependencies that were causing import errors in configuration files.
- typescript-eslint is needed to get those two things working together.
- @eslint/js has the recommended JS linting configuration to use as base for the rest.

Updated dependencies:
- TS 5.1 -> 5.8. Brings Explicit Resource Management, named tuples, groupBy, better inferring and tooling.
- vite 4.4 -> 6.2. Brings boost of performance in rollup, changes in API and improvements all across the board.
- vite-plugin-glsl 1.1.2 -> 1.4. Renamed the 'compress' option to 'minify'.
- ESLint 8.46 -> 9.24. Flatconfig is now the default and well supported (yay!). Improved TS support.